+++ t/chat.t
@@ -32,12 +32,16 @@ $t->post_ok('/login', form => {login => 'doe', password => 'barbar'})->status_is
 
 {
   $t->websocket_ok('/socket');
-  $connection->_irc->from_irc_server(":magnet.llarian.net 353 doe = #convos :fooman \@woman\r\n");
+  $connection->_irc->from_irc_server(
+    ":magnet.llarian.net 353 doe = #convos :fooman {special} \@woman +man [special]\r\n");
   $connection->_irc->from_irc_server(":magnet.llarian.net 366 doe #convos :End of /NAMES list.\r\n");
   $dom->parse($t->message_ok->message->[1]);
   ok $dom->at('li.nick-init[data-network="magnet"][data-target="#convos"]'), 'Got correct names for #convos';
-  is $dom->at('a[href="cmd:///query fooman"]')->text, 'fooman', 'got fooman';
-  is $dom->at('a[href="cmd:///query woman"]')->text,  '@woman', 'got woman';
+  is $dom->at('a[href="cmd:///query fooman"][data-nick="fooman"]')->text,       'fooman',    'got fooman';
+  is $dom->at('a[href="cmd:///query woman"][data-nick="woman"]')->text,         '@woman',    'got woman';
+  is $dom->at('a[href="cmd:///query man"][data-nick="man"]')->text,             '+man',      'got man';
+  is $dom->at('a[href="cmd:///query [special]"][data-nick="[special]"]')->text, '[special]', 'got [special]';
+  is $dom->at('a[href="cmd:///query {special}"][data-nick="{special}"]')->text, '{special}', 'got {special}';
 }
 
 {